journal_entry
I had some issues regarding exercise 3, because I couldn't convert my integer to the long format using the 64-bits, so after a couple of hours I had to look for help online.
For the Fibonacci question I tried to use recursion at first, but that was taking too long for bigger numbers, so I had to change my approach.